Changed session.run not discard previous run result, but buffer all into memory. This enables nested session runs.
