+++ driver/src/main/java/org/neo4j/driver/internal/FailableCursor.java
@@ -22,5 +22,6 @@ import java.util.concurrent.CompletionStage;
 
 public interface FailableCursor
 {
+    CompletionStage<Throwable> consumeAsync();
     CompletionStage<Throwable> failureAsync();
 }
