+++ driver/src/main/java/org/neo4j/driver/internal/async/ResultCursorsHolder.java
@@ -74,6 +74,6 @@ public class ResultCursorsHolder
     {
         return cursorStage
                 .exceptionally( cursor -> null )
-                .thenCompose( cursor -> cursor == null ? completedWithNull() : cursor.failureAsync() );
+                .thenCompose( cursor -> cursor == null ? completedWithNull() : cursor.consumeAsync() );
     }
 }
