+++ driver/src/test/java/org/neo4j/driver/internal/handlers/LegacyPullAllResponseHandlerTest.java
@@ -20,16 +20,12 @@ package org.neo4j.driver.internal.handlers;
 
 import org.junit.jupiter.api.Test;
 
-import java.io.IOException;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.CompletionStage;
 import java.util.function.Function;
 
 import org.neo4j.driver.Record;
 import org.neo4j.driver.Statement;
-import org.neo4j.driver.exceptions.ServiceUnavailableException;
-import org.neo4j.driver.exceptions.SessionExpiredException;
 import org.neo4j.driver.internal.spi.Connection;
 import org.neo4j.driver.summary.ResultSummary;
 
@@ -52,121 +48,6 @@ import static org.neo4j.driver.util.TestUtil.await;
 
 class LegacyPullAllResponseHandlerTest extends PullAllResponseHandlerTestBase<LegacyPullAllResponseHandler>
 {
-    @Test
-    void shouldReturnNoFailureWhenAlreadySucceeded()
-    {
-        LegacyPullAllResponseHandler handler = newHandler();
-        handler.onSuccess( emptyMap() );
-
-        Throwable failure = await( handler.failureAsync() );
-
-        assertNull( failure );
-    }
-
-    @Test
-    void shouldReturnNoFailureWhenSucceededAfterFailureRequested()
-    {
-        LegacyPullAllResponseHandler handler = newHandler();
-
-        CompletableFuture<Throwable> failureFuture = handler.failureAsync().toCompletableFuture();
-        assertFalse( failureFuture.isDone() );
-
-        handler.onSuccess( emptyMap() );
-
-        assertTrue( failureFuture.isDone() );
-        assertNull( await( failureFuture ) );
-    }
-
-    @Test
-    void shouldReturnFailureWhenAlreadyFailed()
-    {
-        LegacyPullAllResponseHandler handler = newHandler();
-
-        RuntimeException failure = new RuntimeException( "Ops" );
-        handler.onFailure( failure );
-
-        Throwable receivedFailure = await( handler.failureAsync() );
-        assertEquals( failure, receivedFailure );
-    }
-
-    @Test
-    void shouldReturnFailureWhenFailedAfterFailureRequested()
-    {
-        LegacyPullAllResponseHandler handler = newHandler();
-
-        CompletableFuture<Throwable> failureFuture = handler.failureAsync().toCompletableFuture();
-        assertFalse( failureFuture.isDone() );
-
-        IOException failure = new IOException( "Broken pipe" );
-        handler.onFailure( failure );
-
-        assertTrue( failureFuture.isDone() );
-        assertEquals( failure, await( failureFuture ) );
-    }
-
-    @Test
-    void shouldReturnFailureWhenRequestedMultipleTimes()
-    {
-        LegacyPullAllResponseHandler handler = newHandler();
-
-        CompletableFuture<Throwable> failureFuture1 = handler.failureAsync().toCompletableFuture();
-        CompletableFuture<Throwable> failureFuture2 = handler.failureAsync().toCompletableFuture();
-
-        assertFalse( failureFuture1.isDone() );
-        assertFalse( failureFuture2.isDone() );
-
-        RuntimeException failure = new RuntimeException( "Unable to contact database" );
-        handler.onFailure( failure );
-
-        assertTrue( failureFuture1.isDone() );
-        assertTrue( failureFuture2.isDone() );
-
-        assertEquals( failure, await( failureFuture1 ) );
-        assertEquals( failure, await( failureFuture2 ) );
-    }
-
-    @Test
-    void shouldReturnFailureOnlyOnceWhenFailedBeforeFailureRequested()
-    {
-        LegacyPullAllResponseHandler handler = newHandler();
-
-        ServiceUnavailableException failure = new ServiceUnavailableException( "Connection terminated" );
-        handler.onFailure( failure );
-
-        assertEquals( failure, await( handler.failureAsync() ) );
-        assertNull( await( handler.failureAsync() ) );
-    }
-
-    @Test
-    void shouldReturnFailureOnlyOnceWhenFailedAfterFailureRequested()
-    {
-        LegacyPullAllResponseHandler handler = newHandler();
-
-        CompletionStage<Throwable> failureFuture = handler.failureAsync();
-
-        SessionExpiredException failure = new SessionExpiredException( "Network unreachable" );
-        handler.onFailure( failure );
-        assertEquals( failure, await( failureFuture ) );
-
-        assertNull( await( handler.failureAsync() ) );
-    }
-
-    @Test
-    void shouldReturnSummaryWhenAlreadyFailedAndFailureConsumed()
-    {
-        Statement statement = new Statement( "CREATE ()" );
-        LegacyPullAllResponseHandler handler = newHandler( statement );
-
-        ServiceUnavailableException failure = new ServiceUnavailableException( "Neo4j unreachable" );
-        handler.onFailure( failure );
-
-        assertEquals( failure, await( handler.failureAsync() ) );
-
-        ResultSummary summary = await( handler.summaryAsync() );
-        assertNotNull( summary );
-        assertEquals( statement, summary.statement() );
-    }
-
     @Test
     void shouldDisableAutoReadWhenTooManyRecordsArrive()
     {
